+++ generators/src/bin/generate_embeddings.rs
@@ -5,7 +5,6 @@ use clap::Parser;
 use rand::SeedableRng;
 use rand_chacha::ChaCha8Rng;
 use serde::{Deserialize, Serialize};
-use serde_json;
 
 #[derive(Serialize, Deserialize)]
 struct Embedding {