+++ vectro_py/src/lib.rs
@@ -1,3 +1,4 @@
+#![allow(non_local_definitions)]
 use pyo3::prelude::*;
 use pyo3::types::{PyList, PyTuple};
 use numpy::{IntoPyArray, PyArray1, PyArray2, PyReadonlyArray1, PyReadonlyArray2};
@@ -148,7 +149,7 @@ impl PySearchIndex {
         let indices_array: &PyArray1<usize> = Array1::from(indices).into_pyarray(py);
         let similarities_array: &PyArray1<f32> = Array1::from(similarities).into_pyarray(py);
         
-        Ok(PyTuple::new(py, &[indices_array.as_ref(), similarities_array.as_ref()]).into())
+        Ok(PyTuple::new(py, [indices_array.as_ref(), similarities_array.as_ref()]).into())
     }
 
     fn batch_search(&self, py: Python<'_>, queries: PyReadonlyArray2<f32>, top_k: usize) -> PyResult<Py<PyList>> {
@@ -171,7 +172,7 @@ impl PySearchIndex {
             
             let indices_array: &PyArray1<usize> = Array1::from(indices).into_pyarray(py);
             let similarities_array: &PyArray1<f32> = Array1::from(similarities).into_pyarray(py);
-            let result_tuple = PyTuple::new(py, &[indices_array.as_ref(), similarities_array.as_ref()]);
+            let result_tuple = PyTuple::new(py, [indices_array.as_ref(), similarities_array.as_ref()]);
             
             all_results.push(result_tuple);
         }
@@ -181,7 +182,7 @@ impl PySearchIndex {
 
     fn __repr__(&self) -> String {
         // We can't access private fields, so use a simpler representation
-        format!("PySearchIndex")
+        "PySearchIndex".to_string()
     }
 }
 
@@ -230,7 +231,7 @@ impl PyQuantizedIndex {
         let indices_array: &PyArray1<usize> = Array1::from(indices).into_pyarray(py);
         let similarities_array: &PyArray1<f32> = Array1::from(similarities).into_pyarray(py);
         
-        Ok(PyTuple::new(py, &[indices_array.as_ref(), similarities_array.as_ref()]).into())
+        Ok(PyTuple::new(py, [indices_array.as_ref(), similarities_array.as_ref()]).into())
     }
 
     fn compression_ratio(&self) -> f32 {